refactor: icons refactor
