+++ packages/error-overlay/src/react/components/Errors/CallStackFrame/CallStackFrame.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import { StackFrame } from 'stacktrace-parser';
+
+import { ExternalLinkIcon } from '../../Icons';
 import {
   getFrameSource,
   OriginalStackFrame
@@ -35,19 +37,7 @@ export const CallStackFrame: React.FC<{
         title={hasSource ? 'Click to open in your editor' : undefined}
       >
         <span>{getFrameSource(f)}</span>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        >
-          <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path>
-          <polyline points="15 3 21 3 21 9"></polyline>
-          <line x1="10" y1="14" x2="21" y2="3"></line>
-        </svg>
+        <ExternalLinkIcon />
       </div>
     </div>
   );
