+++ packages/error-overlay/src/react/components/Errors/CodeFrame/CodeFrame.tsx
@@ -2,6 +2,8 @@ import Anser from 'anser';
 import * as React from 'react';
 import { StackFrame } from 'stacktrace-parser';
 import stripAnsi from 'strip-ansi';
+
+import { ExternalLinkIcon } from '../../Icons';
 import { getFrameSource } from '../../../helpers/stack-frame';
 
 export type CodeFrameProps = { stackFrame: StackFrame; codeFrame: string };
@@ -59,19 +61,7 @@ export const CodeFrame: React.FC<CodeFrameProps> = function CodeFrame({
           <span>
             {getFrameSource(stackFrame)} @ {stackFrame.methodName}
           </span>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path>
-            <polyline points="15 3 21 3 21 9"></polyline>
-            <line x1="10" y1="14" x2="21" y2="3"></line>
-          </svg>
+          <ExternalLinkIcon />
         </p>
       </div>
       <pre>
