+++ packages/error-overlay/src/react/components/Icons/ArrowLeftIcon.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+
+const ArrowLeftIcon = () => {
+  return (
+    <svg viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M6.99996 1.16666L1.16663 6.99999L6.99996 12.8333M12.8333 6.99999H1.99996H12.8333Z"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+
+export { ArrowLeftIcon };
