+++ packages/error-overlay/src/react/components/Icons/CloseIcon.tsx
