+++ packages/error-overlay/src/react/components/Icons/WarningIcon.tsx
