+++ packages/error-overlay/src/react/components/Icons/index.ts
@@ -1,0 +1,5 @@
+export { CloseIcon } from './CloseIcon';
+export { WarningIcon } from './WarningIcon';
+export { ExternalLinkIcon } from './ExternalLinkIcon';
+export { ArrowLeftIcon } from './ArrowLeftIcon';
+export { ArrowRightIcon } from './ArrowRightIcon';
